Added - login end-point

+++ routers/snippetRouter.js
@@ -5,8 +5,8 @@ const auth = require("../middleware/auth");
 //Get Route
 router.get("/", auth, async (req, res) => {
   try {
-   console.log(req.user, "USer");
-    const snippet = await Snippet.find();
+    console.log(req.user, "USer");
+    const snippet = await Snippet.find({user:req.user});
     res.status(200).json(snippet);
   } catch (error) {
     console.log(error);
@@ -18,18 +18,22 @@ router.post("/", auth, async (req, res) => {
   try {
     const { title, description, code } = req.body.snippetInpuptData;
     console.log("POST", title, description, code);
+
     if (!(title && description && code)) {
       return res.status(400).json({ errorMessage: "You need to enter at least description or some code." });
     }
+
     const newSnippet = new Snippet({
       title,
       description,
       code,
+      user: req.user,
     });
+    
     await newSnippet.save();
     res.status(200).json({ message: "Snippet saved successfully" });
   } catch (error) {
-    res.status(500).send();
+    res.status(500).json({errorMessage:"Something went wrong"});
   }
 });
 
@@ -46,6 +50,11 @@ router.delete("/:id", auth, async (req, res) => {
     if (!existingSnippet) {
       res.status(400).json({ errorMessage: "Snippet does not Exist" });
     }
+
+    if(existingSnippet.user.toString() !== req.user){
+      return res.status(401).josn({ errorMessage: "Unauthorized!" });
+    }
+
     await existingSnippet.delete();
     res.json(existingSnippet);
     console.log(snippetId);
@@ -59,19 +68,27 @@ router.put("/:id", auth, async (req, res) => {
   try {
     const { title, description, code } = req.body.snippetInpuptData;
     console.log("PUT", req.body);
+
     if (!(title && description && code)) {
       return res.status(400).json({ errorMessage: "Title, Description, Code can not be empty" });
     }
+
     const snippetId = req.params.id;
     console.log("PUT_ID", snippetId);
+
     if (!snippetId) {
       return res.status(400).json({ errorMessage: "Snippet id not found" });
     }
-
+    
     const originalSnippet = await Snippet.findById(snippetId);
     if (!originalSnippet) {
       return res.status(400).json({ errorMessage: "Snippet dose not Exist" });
     }
+
+    if(originalSnippet.user.toString() !== req.user){
+      return res.status(401).josn({ errorMessage: "Unauthorized!" });
+    }
+
     originalSnippet.title = title;
     originalSnippet.description = description;
     originalSnippet.code = code;
